clang-plugin: Add a second annotater plugin for assertions
This takes assertion conditions and uses them to annotate function
declarations with (currently only) nonnull attributes.

For example, having the following assertion inside a function body will
cause a nonnull attribute to be added for that parameter on the function
declaration:
    g_return_val_if_fail (my_object != NULL, FALSE);

This annotater still needs more support for other attributes, and wider
testing on larger bodies of code, but is a good start.