build: Explicitly check for clang-devel package availability
The Clang development headers are needed to compile the plugin against.
They’re in the clang-devel package on Fedora, or similarly-named
packages on other distributions.

Of course, that package doesn’t install a pkg-config file, so we have to
manually check for the header in the same prefix as LLVM is installed.