clang-plugin: Add support for checking g_variant_get() calls
This required a little further work on top of the work for checking
g_variant_new() calls, as g_variant_get() takes outbound arguments,
rather than inbound ones, so has an extra level of indirection. Further
to that, integer promotion doesn’t occur, and nullable arguments are
treated a little differently.

This includes a set of unit tests for g_variant_get().