clang-plugin: Hard-code GLib functions to ignore for nullability checks
There are several GLib functions which are exported as symbols in
libglib-2.0.so, but which are deliberately not documented or exposed in
the GIR file. They’re for internal use by GLib macros.

Since we can’t extract nullability information from the GIR file for
them, hard-code them into a list of GLib functions to ignore, otherwise
we get loads of false positives (especially in unit test code).