clang-plugin: Automatically promote int to uint with constant varargs
If an integer constant is passed as a vararg, it will be passed as a
signed integer to the function. For integers less than G_MAXINT this is
not a problem (as the bit representations of int and uint are the same
to that point). Modify the GVariant checker to accept low-valued signed
integers for unsigned integer GVariant elements.