tests: Check for all expected error lines rather than just one
Previously, the wrapper-compiler-errors script was assuming a test was
successful if it could see at least one of the expected error message
lines in the compiler output. Since several of those lines are
boilerplate, this hid legitimate test failures.

Fix it so that *all* expected lines are checked for (although there are
no checks which test that *only* those lines were output), and fix up a
few tests.