gsignal: Fix callback instance type checking for supertypes
Previously, Tartan would enforce the type relationship:
    actual <: static
but this was wrong, as then the callback could legitimately be invoked
with an instance of type T, where:
    actual <: T  and  T <: static
and it could proceed to call methods defined on actual which are not
defined on T, and crash.

Instead, the following relationship has to hold:
    static <: actual
which guarantees that for any T passed to the callback, since we have to
have T <: static, we definitely also have T <: actual, which is the
necessary condition for type safety of function parameters on
invocation.

Additionally, check whether
    static = actual
and emit a remark if that relationship does not hold, as the callback is
most likely performing unnecessary and inefficient dynamic type checks
in this case.