gvariant: Check for use of architecture-dependent types
If the user passes an architecture-dependent type into a GVariant
variadic function, the behaviour will differ between architectures,
which we really do not want — essentially the code would be using the
wrong width integer on one architecture but not another, which would be
hard to track down. This is mostly a problem with ‘long’, which is 32
bits wide on 32-bit systems, and 64 bits wide on 64-bit systems.

This was a problem encountered in the wild in telepathy-gabble. Thanks
to Simon McVittie for suggesting that Tartan check for it.