scripts: Handle setting CC=tartan correctly
Previously, if the user mistakenly ran Tartan using `make CC=tartan`, it
would error out, as the Clang wrapper script would be invoked with
-add-plugin not escaped by -Xclang. Normally, when the tartan wrapper
script is used, it is passed a -cc1 option, which causes the clang
compiler (not the wrapper script) to be executed, which expects
-add-plugin rather than -Xclang -add-plugin.

Handle this better by checking for the -cc1 argument and adding -Xclang
as appropriate. Also print a warning in the case that the user has
called `make CC=tartan`, as this will only enable the compile-time
checkers, not the static analyser.

Reported by Marc-Antoine Perennou <Marc-Antoine@Perennou.com>.