gsignal: Allow swapped object instances to have actual type gpointer
It seems common to use a gpointer argument for the swapped instance
parameter (i.e. the last argument) to a signal callback, even though
this is not the most specific type known for the variable.

Allow this behaviour — hopefully the programmer is correctly dynamically
casting the gpointer to something more concrete in the callback body.
Either that, or (more likely), ignoring it entirely.