tests: Update non-null warning messages to support LLVM 3.5
The wording of the non-null warning messages changed between LLVM 3.4
and LLVM 3.5, so the tests were all arbitrarily failing.

This change makes them work on LLVM 3.5 at the expense of failing on
LLVM 3.4. There’s no satisfactory way to improve this, short of
duplicating the test files and having one copy for 3.4 and one for 3.5.