gerror: Fix crash when running on non-GLib code
If the GError checker was run on non-GLib code, the assertion about
having the typedef for the GError structure would fail and cause Clang
to abort. This was commonly being triggered when running the checker in
a configure script.

Fix that by gracefully and rapidly bailing out of the GError checker if
glib.h has not been included in this AST context.