scripts/tartan: improve parsing rule of clang version
clang --version returns version strings, but it varies in
different linux distributions. It would be better to assume
the version string is located in the middle of the string.

For example, clang of debian(jessie) returns like this:

  $ clang --version
  Debian clang version 3.5.0-10 .......

Thus, we cannot assume that the version string is located in
a specific field by "cut -f -d ' '".

https://github.com/pwithnall/tartan/issues/2