scripts: Fix underquoting problem when scan-build has an analyser crash
If the static analyser crashes while analysing a file, scan-build has a
debug mechanism which calls clang in a different mode (-fsyntax-only) to
grab more information about the crash.

This happens through ccc-analyzer calling clang with `-### -fsyntax-only
$args`. This requires the Tartan arguments to be quoted correctly with
-Xclang and -Xanalyzer; they were not being quoted this way. That led to
libtartan.so being chosen as the file to run -fsyntax-only on, resulting
in a terminal full of complaints about nul bytes.

Fix that by always quoting the arguments for -###.

Overall, we want tartan-build to work in the following modes:

In a terminal
        tartan-build jhbuild make                       Yes
        tartan-build jhbuild make -ac
In a JHBuild shell:
        tartan-build autogen.sh; make                   No
        tartan-build autogen.sh; tartan-build make      Yes
        autogen.sh; tartan-build make                   Yes