debug: Emit all errors as warnings
As explained in a comment in the commit, emitting diagnostics as errors
causes scan-build to assume Clang has crashed, and hence abort parsing
that file. That’s not very useful.

On the assumption that a) this should be avoided; and b) Tartan is not
yet mature enough to be confident in emitting diagnostics as definite
errors, change all errors to be emitted as warnings.

The internal API for emitting an error will remain in place to indicate
diagnostics which could potentially be errors in future. These
diagnostics are the ones where there is no conceivable way the code
written by a programmer can be right.